fix(mcts-grpo): add fp16 mixed precision compatibility
When using Accelerate's mixed precision training (fp16) on CUDA, three potential issues could cause numerical instability or dtype mismatches:

- Gradient norm calculation could overflow with fp16 gradients
- Log probability clamp value (1e-10) falls into fp16 denormal range
- init_hidden buffer would not match autocast dtype during forward

These fixes ensure safe training when mixed_precision="fp16" is enabled, while maintaining full backward compatibility with fp32.

+++ server/ml/models/lfu_mcts_grpo/dev/mcts_grpo/model/modeling_urm.py
@@ -386,7 +386,8 @@ class URMModel(PreTrainedModel):
         input_embeddings = self.input_embed(x)  # [batch, 64, hidden_size]
 
         # Initialize hidden state (broadcast to batch)
-        hidden_states = self.init_hidden.expand(
+        # Cast to match input dtype/device for fp16 compatibility
+        hidden_states = self.init_hidden.to(dtype=input_embeddings.dtype).expand(
             batch_size, -1, -1
         )  # [batch, 1, hidden]
         hidden_states = hidden_states.expand(-1, 64, -1)  # [batch, 64, hidden]
